# nodejs-server-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "API handling chaincode calling and identities for Hyperledger Fabric"
  version: "1.0.0"
  title: "Fabric API"
  contact:
    email: "apiteam@swagger.io"
...
paths:
...
  /users/{username}/{projectid}/submitreview:
    post:
      tags:
      - "Project management"
      summary: "allow reviewers to submit their review on a project"
      description: ""
      operationId: "SubmitReview"
      produces:
      - "application/json"
      - "text/plain"
      parameters:
      - name: "X-request-use-token"
        in: "header"
        description: "boolean, true if a session token is used"
        required: false
        type: "boolean"
      - name: "X-request-token"
        in: "header"
        description: "session token"
        required: false
        type: "string"
      - name: "username"
        in: "path"
        description: "Username"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-request-password"
        in: "header"
        description: "Password of the user"
        required: false
        type: "string"
        format: "string"
      - name: "X-request-review"
        in: "header"
        description: "Review, json format"
        required: true
        type: "string"
        format: "string"
      - name: "projectid"
        in: "path"
        description: "projectid where to add a reviewer"
        required: true
        type: "string"
        format: "string"
 
      responses:
        200:
          description: "Reviewer link succesfully generated"
          schema:
            $ref: "#/definitions/inline_response_200"
        401:
          description: "wrong username/pw or wrong project"
          schema:
            type: "string"
        404:
          description: "username not found"
          schema:
            type: "string"
      x-swagger-router-controller: "SubmitReview"
...
  /users/{username}/validate/{token}:
    get:
      tags:
      - "Identity management"
      summary: "confirm account via a link sent by e-mail"
      description: ""
      operationId: "Validate"
      produces:
      - "application/json"
      - "text/plain"
      parameters:
        - name: "username"
          in: "path"
          description: "Username of the user validating his account"
          required: true
          type: "string"
          format: "uuid"
        - name: "token"
          in: "path"
          description: "unique token sent by email"
          required: true
          type: "string"
          format: "uuid"
      responses:
        200:
          description: "User misc data successfully retreived"
          schema:
            $ref: "#/definitions/inline_response_200"
        404:
          description: "username not found"
          schema:
            type: "string"
      x-swagger-router-controller: "Validate"
...
definitions:
...
  inline_response_200:
    properties:
      pubkey:
        type: "string"
        description: "The public key of the given user"
    example:
      pubkey: "pubkey"
...